Enhancement: Decompose CompactUtils.ts into focused modules (#1745)
Split CompactUtils.ts (618 lines, 4 responsibilities) into single-responsibility modules:
- OrphanedNeuronCleanup.ts — orphaned neuron detection and removal
- SynapsePruning.ts — duplicate and zero-weight synapse pruning
- DeadSubgraphPruning.ts — unreachable subgraph removal
- MemeticCleanup.ts — memetic data cleanup for removed structures

CompactUtils.ts is now a thin re-export barrel file for backward compatibility.
Pure refactoring — no functional changes.
